Refactored cards and token components
