+++ src/components/card/SelectCardModal.tsx
@@ -1,20 +1,15 @@
-import { Box, Flex, Input, Text } from "@chakra-ui/react";
+import { Flex} from "@chakra-ui/react";
 import { useRecoilState } from "recoil";
-
 import { Modal, ModalOverlay, ModalContent, ModalBody } from "@chakra-ui/react";
-import { useCallback, useEffect, useRef, useState } from "react";
+import { useCallback, useEffect, useRef } from "react";
 import Image from "next/image";
 import useTokenModal from "@/hooks/modal/useTokenModal";
-import { Field } from "@/types/swap/swap";
 import { CardCarrousel } from "./CardCarousel";
 import {
-  searchTokenStatus,
   IsSearchToken,
   isInputTokenAmount,
   isOutputTokenAmount,
 } from "@/recoil/card/selectCard/searchToken";
-import useConnectedNetwork from "@/hooks/network";
-import { Overlay_Index } from "@/types/style/overlayIndex";
 import { CardCarouselMobile } from "./mobile/CardCarouselMobile";
 import useMediaView from "@/hooks/mediaView/useMediaView";
 import TokenInput from "../input/TokenInput";
@@ -23,153 +18,9 @@ import {
   tokenModalStatus,
 } from "@/recoil/bridgeSwap/atom";
 import { useRecoilValue } from "recoil";
-
 import BgImage from "assets/image/BridgeSwap/selectTokenCardBg.svg";
-import BgImageButton from "assets/image/BridgeSwap/selectTokenBg.svg";
-import CloseIcon from "assets/icons/close.svg";
-import SearchIcon from "assets/icons/search.svg";
-import CancelIcon from "assets/icons/close.svg";
-import { isIOS } from 'react-device-detect';
-
-export function SelectCardButton(props: { field: Field }) {
-  const { field } = props;
-  const { onOpenInToken, onOpenOutToken } = useTokenModal();
-
-  return (
-    <Flex
-      w={"562px"}
-      h={"100px"}
-      alignItems={"center"}
-      justifyContent={"center"}
-      cursor={"pointer"}
-      onClick={() => (field === "INPUT" ? onOpenInToken() : onOpenOutToken())}
-      pos={"relative"}
-    // zIndex={Overlay_Index}
-    >
-      <Image
-        src={BgImageButton}
-        alt={"BgImageButton"}
-        style={{ position: "absolute" }}
-      />
-      <Text
-        color={"#FFFFFF"}
-        fontSize={24}
-        fontWeight={"semibold"}
-        zIndex={100}
-        mt={"10px"}
-      >
-        Select Token
-      </Text>
-    </Flex>
-  );
-}
-
-const SearchToken = () => {
-  const { onCloseTokenModal } = useTokenModal();
-  const [, setSearchToken] = useRecoilState(searchTokenStatus);
-  const { mobileView, pcView } = useMediaView();
-
-  const { connectedChainId } = useConnectedNetwork();
-  const ref = useRef<HTMLInputElement>(null);
-  const [searchValue, setSearchValue] = useState<string>("");
-  const [, setTokenSearch] = useRecoilState(IsSearchToken);
-  const [isInputAmount, setIsInputAmount] = useRecoilState(isInputTokenAmount);
-  const [selectedInToken] = useRecoilState(selectedInTokenStatus);
-
-  useEffect(() => {
-    setTimeout(() => {
-      ref.current?.blur();
-      if (selectedInToken?.amountBN) {
-        setIsInputAmount(true);
-      }
-    }, 20);
-  }, []);
-
-  const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    e.preventDefault();
-    const value = e.target.value;
-    setSearchValue(value);
-  };
-
-  const handleFocus = () => {
-    setTokenSearch(true);
-    setIsInputAmount(false);
-  }
-
-  const handleBlur = () => {
-    setTokenSearch(false);
-  }
-
-  const onKeyDown = (e: any) => {
-    if (e.key === "Enter" && mobileView) {
-      ref?.current?.blur();
-    }
-  }
-
-  useEffect(() => {
-    if (searchValue === "") {
-      return setSearchToken(null);
-    }
-    if (connectedChainId) {
-      return setSearchToken({ nameOrAdd: searchValue, chainId: connectedChainId });
-    }
-  }, [searchValue])
-
-  return (
-    <Flex
-      w={"100%"}
-      justifyContent={"center"}
-      pos={"relative"}
-      zIndex={Overlay_Index.BelowHeader}
-      border={"1px solid transparent"}
-      _hover={{ border: mobileView ? "1px solid #313442" : "" }}
-      rounded={{ base: "8px", lg: "21.5px" }}
-      bgColor={{ base: "#0F0F12", lg: "transparent" }}
-    >
-      <Input
-        w={{ base: "100%", lg: "430px" }}
-        h={"42px"}
-        borderRadius={{ base: "8px", lg: "21.5px" }}
-        placeholder={"Search token name or address"}
-        _placeholder={{ color: "#8E8E92", fontWeight: 500 }}
-        boxShadow={"none !important"}
-        border={{}}
-        bgColor={"#0F0F12"}
-        _focus={{}}
-        _active={{}}
-        onChange={onChange}
-        onFocus={handleFocus}
-        onBlur={handleBlur}
-        onKeyDown={onKeyDown}
-        ref={ref}
-        value={searchValue}
-      ></Input>
-
-      {mobileView && (
-        <Image
-          src={searchValue ? CancelIcon : SearchIcon}
-          alt={"close"}
-          style={{ cursor: "pointer", marginRight: "10px" }}
-          onMouseDown={(e) => {
-            e.preventDefault();
-          }}
-          onClick={() => { setSearchValue("") }}
-        />
-      )}
-
-      {pcView && (
-        <Box pos={"absolute"} right={"69px"}>
-          <Image
-            src={CloseIcon}
-            alt={"close"}
-            style={{ cursor: "pointer" }}
-            onClick={() => onCloseTokenModal()}
-          />
-        </Box>
-      )}
-    </Flex>
-  );
-};
+import { isIOS } from "react-device-detect";
+import { SearchCardToken } from "./SearchCardToken";
 
 export function SelectCardModal() {
   const { isInTokenOpen, isOutTokenOpen, onCloseTokenModal } = useTokenModal();
@@ -222,7 +73,10 @@ export function SelectCardModal() {
         minW={"100%"}
         maxW={"100%"}
         h={{
-          base: (isInputAmount && !isIOS || isTokenSearch && !isIOS) ? "calc(100% - 60px)" : "fit-content",
+          base:
+            (isInputAmount && !isIOS) || (isTokenSearch && !isIOS)
+              ? "calc(100% - 60px)"
+              : "fit-content",
           lg: "100%",
         }}
         m={{ base: "none", lg: 0 }}
@@ -234,7 +88,7 @@ export function SelectCardModal() {
         bg={{ base: "#1F2128", lg: "transparent" }}
         overflow={"hidden"}
         borderRadius={"24px 24px 0px 0px"}
-        >
+      >
         <ModalBody
           minW={"100%"}
           maxW={"100%"}
@@ -245,7 +99,7 @@ export function SelectCardModal() {
           bg={"transparent"}
           id="out-area"
           zIndex={1}
-          >
+        >
           <Flex
             w={"1362px"}
             h={{ base: "100%", lg: "486px" }}
@@ -277,34 +131,41 @@ export function SelectCardModal() {
             {pcView && (
               <>
                 <CardCarrousel />
-                <SearchToken />
+                <SearchCardToken />
               </>
             )}
             {!pcView && (
               <>
-                <SearchToken />
+                <SearchCardToken />
                 <CardCarouselMobile />
                 <Flex
                   w={"full"}
                   justify={"center"}
                   align={"start"}
                   columnGap={"11px"}
-                // onBlur={handleBlur}
-                // px={"10px"}
+                  // onBlur={handleBlur}
+                  // px={"10px"}
                 >
-                  {((isInputAmount && isInTokenOpen) || (isOutputAmount && isOutTokenOpen)) ? 
+                  {(isInputAmount && isInTokenOpen) ||
+                  (isOutputAmount && isOutTokenOpen) ? (
                     <TokenInput
                       inToken={isOpen === "INPUT" ? true : false}
                       hasMaxButton={isOpen === "INPUT" ? true : false}
-                      style={isInputAmount || isOutputAmount ? "" : { display: "none" }}
+                      style={
+                        isInputAmount || isOutputAmount
+                          ? ""
+                          : { display: "none" }
+                      }
                       customRef={ref}
                       placeholder={"input amount"}
                       isDisabled={isOpen === "INPUT" ? false : true}
                       defaultValue={
                         isOpen === "INPUT" ? selectedInToken?.parsedAmount : ""
                       }
-                    /> : <></>
-                  }
+                    />
+                  ) : (
+                    <></>
+                  )}
                   {/* <Flex
                     minW={"40px"}
                     minH={"40px"}
